get rid of int dtype warnings

+++ tests/test_timeseriesx.py
@@ -25,7 +25,7 @@ def empty_timestamp_series():
 @pytest.fixture
 def default_timestamp_series(n=3):
     return TimestampSeries(
-        pd.Series(PintArray(np.arange(n), dtype=ureg.parse_units('m')),
+        pd.Series(PintArray(np.arange(n, dtype=float), dtype=ureg.parse_units('m')),
                   index=pd.date_range('2020-01-01', freq='D', periods=n, tz='CET')
                   ),
         time_zone='CET', unit='m', freq='D'
@@ -48,7 +48,7 @@ def test_timestamp_series_create_from_lists():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [-1, 0, 1]
+    values = [-1., 0., 1.]
     ts = TimestampSeries.create_from_lists(timestamps, values)
     assert ts.freq == pd.offsets.Hour()
     assert ts.time_zone is None
@@ -60,7 +60,7 @@ def test_timestamp_series_create_from_lists_mismatching_length():
         dt.datetime(2020, 3, 1, 15, 0, 0),
         dt.datetime(2020, 3, 1, 16, 0, 0),
     ]
-    values = [-1]
+    values = [-1.]
     with pytest.raises(ValueError):
         TimestampSeries.create_from_lists(timestamps, values)
 
@@ -191,7 +191,7 @@ def test_timestamp_series_as_pd_series_default(default_timestamp_series):
 def test_timestamp_series_as_pd_series_empty(empty_timestamp_series):
     pd.testing.assert_series_equal(
         empty_timestamp_series.as_pd_series(),
-        pd.Series([], index=pd.DatetimeIndex([])),
+        pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
     )
 
 
@@ -365,7 +365,7 @@ def test_timestamp_series_append_no_freq():
 
 
 def test_timestamp_series_prepend_default(default_timestamp_series):
-    assert default_timestamp_series.prepend(-1).first == (
+    assert default_timestamp_series.prepend(-1.).first == (
         pd.Timestamp('2019-12-31T00:00:00').tz_localize('CET').to_pydatetime(), -1.
     )
 
@@ -539,56 +539,56 @@ def test_create_timestamp_series_invalid_index():
 
 
 def test_create_timestamp_series_valid_unit_str():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          unit='second')
     assert ts.unit == ureg.second
 
 
 def test_create_timestamp_series_valid_unit_obj():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          unit=ureg.second)
     assert ts.unit == ureg.second
 
 
 def test_create_timestamp_series_valid_freq_str():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          freq='30Min')
     assert ts.freq == pd.offsets.Minute(30)
 
 
 def test_create_timestamp_series_valid_freq_timedelta():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          freq=pd.Timedelta(minutes=30))
     assert ts.freq == pd.offsets.Minute(30)
 
 
 def test_create_timestamp_series_valid_freq_offset():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          freq=pd.offsets.Minute(30))
     assert ts.freq == pd.offsets.Minute(30)
 
 
 def test_create_timestamp_series_valid_timezone_str():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          time_zone='utc')
     assert ts.time_zone == pytz.UTC
 
 
 def test_create_timestamp_series_valid_timezone_obj_pytz():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          time_zone=pytz.UTC)
     assert ts.time_zone == pytz.UTC
 
 
 def test_create_timestamp_series_valid_timezone_obj_dateutil():
-    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([])),
+    ts = TimestampSeries(series=pd.Series([], index=pd.DatetimeIndex([]), dtype=np.float64),
                          time_zone=dateutil.tz.tzutc())
     assert isinstance(ts.time_zone, tzutc)
 
 
 def test_timestamp_series_add_timestamp_series_different_freq(default_timestamp_series):
     add_ts = TimestampSeries(
-        pd.Series(PintArray(np.arange(3), dtype=ureg.parse_units('m')),
+        pd.Series(PintArray(np.arange(3, dtype=float), dtype=ureg.parse_units('m')),
                   index=pd.date_range('2020-01-01', freq='H', periods=3, tz='CET')
                   ),
         time_zone='CET', unit='m', freq='H'
@@ -599,7 +599,7 @@ def test_timestamp_series_add_timestamp_series_different_freq(default_timestamp_
 
 def test_timestamp_series_add_timestamp_series_different_tz(default_timestamp_series):
     add_ts = TimestampSeries(
-        pd.Series(PintArray(np.arange(3), dtype=ureg.parse_units('m')),
+        pd.Series(PintArray(np.arange(3, dtype=float), dtype=ureg.parse_units('m')),
                   index=pd.date_range('2020-01-01', freq='D', periods=3, tz='UTC')
                   ),
         time_zone='UTC', unit='m', freq='D'
@@ -614,7 +614,7 @@ def test_timestamp_series_add_timestamp_series_different_tz(default_timestamp_se
 
 def test_timestamp_series_add_timestamp_series_different_unit(default_timestamp_series):
     add_ts = TimestampSeries(
-        pd.Series(PintArray(np.arange(3), dtype=ureg.parse_units('kg')),
+        pd.Series(PintArray(np.arange(3, dtype=float), dtype=ureg.parse_units('kg')),
                   index=pd.date_range('2020-01-01', freq='D', periods=3, tz='CET')
                   ),
         time_zone='CET', unit='kg', freq='D'
@@ -625,7 +625,7 @@ def test_timestamp_series_add_timestamp_series_different_unit(default_timestamp_
 
 def test_timestamp_series_add_timestamp_series_different_index(default_timestamp_series):
     add_ts = TimestampSeries(
-        pd.Series(PintArray(np.arange(3), dtype=ureg.parse_units('m')),
+        pd.Series(PintArray(np.arange(3, dtype=float), dtype=ureg.parse_units('m')),
                   index=pd.date_range('2020-01-02', freq='D', periods=3, tz='CET')
                   ),
         time_zone='CET', unit='m', freq='D'
@@ -642,7 +642,7 @@ def test_timestamp_series_add_timestamp_series_different_index(default_timestamp
 
 def test_timestamp_series_add_pandas_series():
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3, tz='UTC')
                   ),
         time_zone='UTC', freq='D'
@@ -662,7 +662,7 @@ def test_timestamp_series_add_pandas_series():
 
 def test_timestamp_series_add_pandas_series_different_time_zones():
     ts = TimestampSeries(
-        pd.Series(np.arange(3.),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3, tz='UTC')
                   ),
         time_zone='Europe/Stockholm', freq='D'
@@ -689,7 +689,7 @@ def test_timestamp_series_add_pandas_series_different_time_zones():
 
 def test_timestamp_series_add_list_error():
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3, tz='UTC')
                   ),
         time_zone='UTC', freq='D'
@@ -723,7 +723,7 @@ def test_timestamp_series_multiply_pint_array(default_timestamp_series):
 
 def test_timestamp_series_add_scalar():
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3, tz='UTC')
                   ),
         time_zone='UTC', freq='D'
@@ -739,7 +739,7 @@ def test_timestamp_series_multiply_scalar(default_timestamp_series):
 
 def test_timestamp_series_multiply_pint_scalar(default_timestamp_series):
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3)
                   ),
         unit='m'
@@ -761,7 +761,7 @@ def test_timestamp_series_floordiv_pint_scalar(default_timestamp_series):
 
 def test_timestamp_series_div_pint_scalar():
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3)
                   ),
         unit='m^2'
@@ -773,7 +773,7 @@ def test_timestamp_series_div_pint_scalar():
 
 def test_timestamp_series_subtract_list():
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3)
                   ))
     result_ts = ts - [0., 1., 2.]
@@ -782,7 +782,7 @@ def test_timestamp_series_subtract_list():
 
 def test_timestamp_series_subtract_pd_series_mismatch():
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3)),
     )
     result_ts = ts - pd.Series(np.arange(2),
@@ -881,7 +881,7 @@ def test_neq_4(default_timestamp_series):
 
 def test_timestamp_series_convert_unit_from_none():
     ts = TimestampSeries(
-        pd.Series(np.arange(3),
+        pd.Series(np.arange(3, dtype=float),
                   index=pd.date_range('2020-01-01', freq='D', periods=3, tz='UTC')
                   ),
         time_zone='UTC', freq='D'
@@ -916,7 +916,7 @@ def test_timestamp_series_fill_gaps():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, 1]
+    values = [1., 1., 1.]
     ts = TimestampSeries.create_from_lists(timestamps, values, freq=pd.offsets.Hour(),
                                            time_zone='Europe/Berlin')
     ts = ts.fill_gaps(ts.start - dt.timedelta(hours=1),
@@ -934,7 +934,7 @@ def test_timestamp_series_fill_gaps_start_and_end_different_timezone():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, 1]
+    values = [1., 1., 1.]
     ts = TimestampSeries.create_from_lists(
         timestamps,
         values,
@@ -955,7 +955,7 @@ def test_timestamp_series_fill_gaps_start_different_timezone():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, 1]
+    values = [1., 1., 1.]
     ts = TimestampSeries.create_from_lists(
         timestamps,
         values,
@@ -975,7 +975,7 @@ def test_timestamp_series_fill_gaps_start_and_end_no_timezone():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, 1]
+    values = [1., 1., 1.]
     ts = TimestampSeries.create_from_lists(
         timestamps,
         values,
@@ -998,12 +998,12 @@ def test_timestamp_series_fill_gaps_no_start_and_end():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, np.nan]
+    values = [1., 1., np.nan]
     ts = TimestampSeries.create_from_lists(timestamps, values, freq=pd.offsets.Hour())
     ts = ts.fill_gaps(value=2)
     assert len(ts) == 3
     assert ts.timestamps == timestamps
-    assert ts.values == [1, 1, 2]
+    assert ts.values == [1., 1., 2.]
 
 
 def test_timestamp_series_fill_gaps_with_unit(default_timestamp_series):
@@ -1026,7 +1026,7 @@ def test_timestamp_series_get_gaps():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, 1]
+    values = [1., 1., 1.]
     ts = TimestampSeries.create_from_lists(timestamps, values, freq=pd.offsets.Hour())
     gaps = ts.get_gaps(timestamps[0] - dt.timedelta(hours=1),
                        timestamps[-1] + dt.timedelta(hours=1))
@@ -1042,7 +1042,7 @@ def test_timestamp_series_get_gaps_start_and_end_different_timezone():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, 1]
+    values = [1., 1., 1.]
     ts = TimestampSeries.create_from_lists(
         timestamps,
         values,
@@ -1062,7 +1062,7 @@ def test_timestamp_series_get_gaps_start_and_end_no_timezone():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [1, 1, 1]
+    values = [1., 1., 1.]
     ts = TimestampSeries.create_from_lists(
         timestamps,
         values,
@@ -1087,7 +1087,7 @@ def test_timestamp_series_get_gaps_no_start_and_end():
         dt.datetime(2020, 3, 1, 16, 0, 0),
         dt.datetime(2020, 3, 1, 17, 0, 0),
     ]
-    values = [np.nan, 1, 1]
+    values = [np.nan, 1., 1.]
     ts = TimestampSeries.create_from_lists(timestamps, values, freq=pd.offsets.Hour())
     gaps = ts.get_gaps()
     assert gaps == [
@@ -1117,7 +1117,7 @@ def test_timestamp_series_resample_without_unit():
     )
     ts = ts.resample('D', 'sum')
     assert ts.freq == pd.offsets.Day()
-    assert ts.values == [24, 24]
+    assert ts.values == [24., 24.]
     assert ts.unit == None
 
 
@@ -1130,7 +1130,7 @@ def test_timestamp_series_resample_str_method():
     )
     ts = ts.resample('D', 'sum')
     assert ts.freq == pd.offsets.Day()
-    assert ts.values == [24, 24]
+    assert ts.values == [24., 24.]
     assert ts.unit == ureg.Unit('m')
 
 
@@ -1143,7 +1143,7 @@ def test_timestamp_series_resample_missing_values():
     )
     ts = ts.resample('D', 'sum')
     assert ts.freq == pd.offsets.Day()
-    assert ts.values == [12]
+    assert ts.values == [12.]
     assert ts.unit == ureg.Unit('m')
